Adding beta to master

+++ source/Axolotl1.1/data.cpp
@@ -1,9 +1,11 @@
 #include "data.h"
 #include "ui_data.h"
-#include "/home/nvidia/wombat/source/source/daemon_manager_src.cpp"
+#include "/home/nvidia/wombat/source/data_logging/authentication.cpp"
 #include <QString>
 #include<QInputDialog>
 #include<signal.h>
+#include <fstream>
+#include <iostream>
 
 Data::Data(QWidget *parent, pid_t a) :
     QWidget(parent),
@@ -22,37 +24,65 @@ void Data::on_pushButton_clicked()
 {
     QString in;
     bool yes;
-    in = QInputDialog::getText(0, "Challenge","Password:", QLineEdit::Normal,"", &yes);
-    std::string pass = in.toUtf8().constData();
 
-    string truekey = "";
-    ifstream truekeyf;
-    string hashfilePath = "/home/nvidia/wombat/source/hashkey"; //normally runDirectory + "/hashkey"
-    truekeyf.open(hashfilePath);
-    if(truekeyf.is_open()) {
-      getline(truekeyf,truekey);
-    }
-    truekeyf.close();
-
-    if(axolotlFileSystem::hash(pass) == truekey){
-        kill(dmid,SIGUSR1);
-    }
-    else{
-        std::cout<<truekey<<" get good\n";
+    // Check if the vehicle is moving, and prevent options from appearing if it is
+    ifstream s_file;
+    s_file.open("/home/nvidia/axolotl/vspeed");
+    int vspeed = 0;
+    if(s_file.is_open()) {
+      std::string vspeed_str;
+      s_file >> vspeed_str;
+      s_file.close();
+      vspeed = atoi(vspeed_str.c_str());
     }
 
+    if(vspeed == 0) {
+      system("dbus-send --type=method_call --dest=org.onboard.Onboard /org/onboard/Onboard/Keyboard org.onboard.Onboard.Keyboard.Show");
+      in = QInputDialog::getText(0, "Clear Data","Enter Password:", QLineEdit::Normal,"", &yes);
+      system("dbus-send --type=method_call --dest=org.onboard.Onboard /org/onboard/Onboard/Keyboard org.onboard.Onboard.Keyboard.Hide");
+      std::string pass = in.toUtf8().constData();
+
+      string truekey = "";
+      ifstream truekeyf;
+      //string hashfilePath = "/home/nvidia/wombat/source/data_logging/hashkey"; //normally runDirectory + "/hashkey"
+      string hashfilePath = "/home/nvidia/axolotl/hashkey"; //normally runDirectory + "/hashkey"
+      truekeyf.open(hashfilePath);
+      if(truekeyf.is_open()) {
+        getline(truekeyf,truekey);
+      }
+      truekeyf.close();
+      if(axolotlFileSystem::hash(pass) == truekey){
+          kill(dmid,SIGUSR1);
+      }
+      else{
+          std::cout << "Password Incorrect: " << truekey << std::endl;
+      }
+    }
 }
 
 void Data::on_pushButton_2_clicked()
 {
     QString in;
     bool yes;
-    in = QInputDialog::getText(0, "Challenge"," Old Password:", QLineEdit::Normal,"", &yes);
+    system("dbus-send --type=method_call --dest=org.onboard.Onboard /org/onboard/Onboard/Keyboard org.onboard.Onboard.Keyboard.Show");
+    in = QInputDialog::getText(0, "Change Password","Current Password:", QLineEdit::Normal,"", &yes);
     std::string oldpass = in.toUtf8().constData();
-    in = QInputDialog::getText(0, "Challenge"," New Password:", QLineEdit::Normal,"", &yes);
 
-    std::string newpass = in.toUtf8().constData();
-    changePassword(oldpass,newpass);
+    string truekey = "";
+    ifstream truekeyf;
+    //string hashfilePath = "/home/nvidia/wombat/source/data_logging/hashkey"; //normally runDirectory + "/hashkey"
+    string hashfilePath = "/home/nvidia/axolotl/hashkey"; //normally runDirectory + "/hashkey"
+    truekeyf.open(hashfilePath);
+    if(truekeyf.is_open()) {
+      getline(truekeyf,truekey);
+    }
+    truekeyf.close();
+    if(axolotlFileSystem::hash(oldpass) == truekey){
+        in = QInputDialog::getText(0, "Change Password","New Password:", QLineEdit::Normal,"", &yes);
+        std::string newpass = in.toUtf8().constData();
+        changePassword(oldpass,newpass);
+    }
+    system("dbus-send --type=method_call --dest=org.onboard.Onboard /org/onboard/Onboard/Keyboard org.onboard.Onboard.Keyboard.Hide");
 }
 
 void Data::on_pushButton_3_clicked()
